+++ src/components/ConnectWallet/ConnectWalletDialog.tsx
@@ -1,13 +1,13 @@
 import { Trans } from '@lingui/macro'
 import { Web3ReactHooks } from '@web3-react/core'
-import { WalletConnect } from '@web3-react/walletconnect'
+import { URI_AVAILABLE, WalletConnect } from '@web3-react/walletconnect'
 import METAMASK_ICON_URL from 'assets/images/metamaskIcon.png'
 import WALLETCONNECT_ICON_URL from 'assets/images/walletConnectIcon.svg'
 import Button from 'components/Button'
 import Column from 'components/Column'
 import { Header } from 'components/Dialog'
 import Row from 'components/Row'
-import useConnect, { connections, Web3Connection } from 'hooks/connectWeb3/useConnect'
+import { connections, useConnect, Web3Connection } from 'hooks/connectWeb3/useActiveWeb3React'
 import { atom, useAtom } from 'jotai'
 import QRCode from 'qrcode'
 import { useEffect, useState } from 'react'
@@ -73,15 +73,18 @@ function WalletConnectButton({ walletName, logoSrc, caption, connection: wcTileC
     if (QRUri) {
       formatQrCodeImage(QRUri)
     } else {
+      console.log('activates')
       tileConnector.activate()
     }
+    // feel like this def gets used in lots of other places, how do other dapps handle this?
+
     // FIX: handle on error/just recall
     // error: if we're connected, then reload the page -- we get error  POST https://mainnet.infura.io/v3/undefined 401
     // at what point does the same QR code URI expire?
     // error: if we close the popup modal before we connect, error
     return () => {
       console.log('remove event listener')
-      // ;(tileConnector.provider?.connector as unknown as EventEmitter | undefined)?.off('display_uri', this.URIListener)
+      // ;(tileConnector.provider?.connector as unknown as EventEmitter | undefined)?.off('display_uri', handleDisplayUri)
     }
   }, [QRUri, tileConnector])
 
@@ -89,9 +92,16 @@ function WalletConnectButton({ walletName, logoSrc, caption, connection: wcTileC
     if (err) console.warn(err)
     // Clear saved QR URI after disconnection
     setQRUri(undefined)
+    tileConnector.deactivate()
+  })
+
+  // log URI when available
+  tileConnector.events.on(URI_AVAILABLE, (uri: string) => {
+    console.log(`we have uri: ${uri}`)
   })
 
   tileConnector.provider?.connector.on('display_uri', async (err, payload) => {
+    console.log('displayuri')
     if (err) console.warn(err)
     const uri: string = payload.params[0]
     if (uri) {
